LayoutTests/imported/w3c/web-platform-tests/shadow-dom/form-control-form-attribute-expected.txt

form


PASS Form control's form attribute should point to the form element.
FAIL Shadow form control's form attribute should work also in shadow DOM. assert_equals: expected null but got Element node <form id="form">form</form>
PASS Form element as form control's ancestor should work also in shadow DOM.
